Add MoG noise model

+++ utils/read_data.py
@@ -6,7 +6,7 @@ def read_s_dataset(s):
     X = pd.read_csv(
         f'data/s-originals/s{s}.txt', header=None, sep='   ', engine='python', dtype=np.float32)
     y = pd.read_csv(
-        f'data/s-originals/s{s}-label.pa', header=None, sep='   ', skiprows=5, engine='python', dtype=np.float32)[0]
+        f'data/s-originals/s{s}-label.pa', header=None, sep='   ', skiprows=5, engine='python', dtype=np.int32)[0]
     return X.values, y.values
 
 
@@ -14,5 +14,5 @@ def read_letter_dataset():
     X = pd.read_csv(f'data/letter/letter.txt',
                     header=None, sep=' ', dtype=np.float32)
     y = pd.read_csv('data/letter/letter.pa', header=None, sep='   ',
-                    skiprows=5, engine='python', dtype=np.float32)[0]
+                    skiprows=5, engine='python', dtype=np.int32)[0]
     return X.values, y.values
